update autofill for add_record_fragment
