+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".storageActivity"
+            android:name=".storage.storageActivity"
             android:exported="false"/>
         <activity
             android:name=".MainActivity"
