+++ app/src/main/java/com/example/fishbook/MainActivity.kt
@@ -4,8 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
-import com.example.fishbook.fishdex.DataRepository
-import com.example.fishbook.gallery.ImageRepository
+import com.example.fishbook.storage.DataRepository
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
