+++ app/src/main/java/com/example/fishbook/fishdex/SpeciesDetailViewModel.kt
@@ -3,6 +3,7 @@ package com.example.fishbook.fishdex
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
+import com.example.fishbook.storage.DataRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
