+++ app/src/main/java/com/example/fishbook/storage/Application.kt
@@ -1,0 +1,13 @@
+package com.example.fishbook.storage
+
+import android.app.Application
+import com.google.firebase.FirebaseApp
+
+class Application : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        DataRepository.initialize(this)
+        FirebaseApp.initializeApp(this)
+
+    }
+}
