+++ app/src/main/res/layout/activity_storage.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:paddingHorizontal="16dp"
-    tools:context=".storageActivity">
+    tools:context=".storage.storageActivity">
 
     <TextView
         android:layout_width="match_parent"
