+++ app/src/main/res/layout/fragment_add_record.xml
@@ -57,16 +57,19 @@
             android:hint="@string/lure_hint"
             android:inputType="text"/>
 
-        <EditText
+        <AutoCompleteTextView
             android:id="@+id/lake_edit_text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_margin="10dp"
             android:layout_below="@id/species_edit_text"
+            android:layout_marginStart="10dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginEnd="10dp"
+            android:layout_marginBottom="10dp"
             android:hint="@string/lake_hint"
-            android:inputType="text"/>
+            android:inputType="text" />
 
-        <EditText
+        <AutoCompleteTextView
             android:id="@+id/county_edit_text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
